Corrige cast Int32/Int64 en lectura de documentos y unifica label de log Padron Socios

+++ Data/AppDbContext.cs
@@ -298,7 +298,7 @@ namespace Implementador.Data
 
                 using var reader = command.ExecuteReader();
                 while (reader.Read())
-                    resultado.Add(reader.GetInt64(0));
+                    resultado.Add((long)reader.GetInt32(0));
             }
 
             return resultado;
